fix(feature-factory): harden collection spec boundaries from review

+++ feature/feature-factory/src/main/java/com/ghealth/tools/feature/factory/engine/TestRawDataCollector.kt
@@ -52,7 +52,7 @@ class TestRawDataCollector @Inject constructor(
     /** 采集是否满足完成条件：去重帧数 >= skip+min，且要求连续时末尾连续帧数 >= min。 */
     fun isCollectionComplete(spec: CollectionSpec): Boolean {
         synchronized(lock) {
-            if (buffers.frameCount() < spec.skipNumber + spec.minNumber) return false
+            if (buffers.frameCount().toLong() < spec.skipNumber.toLong() + spec.minNumber.toLong()) return false
             if (!spec.isContinuous) return true
             return buffers.lastConsecutiveCount() >= spec.minNumber
         }
